fix(runtime-tags): report readonly compile error for <id>/<define> tag variable assignments
