+++ packages/runtime-tags/src/__tests__/fixtures/error-assign-to-define-var/template.marko
@@ -1,0 +1,2 @@
+<define/greeting value="hello"/>
+<button onClick() { greeting = { value: "hi" } }>${greeting.value}</button>
