+++ packages/runtime-tags/src/__tests__/fixtures/error-assign-to-id-var/template.marko
@@ -1,0 +1,2 @@
+<id/uid/>
+<button onClick() { uid = "changed" }>${uid}</button>
