+++ packages/runtime-tags/src/__tests__/fixtures/error-assign-to-id-var/test.ts
@@ -1,0 +1,5 @@
+import type { TestConfig } from "../../main.test";
+
+export const config: TestConfig = {
+  error_compiler: true,
+};
